Move DiagnosePatientFragment to Fragments
